Block Bindings: Componentize
