+++ packages/block-editor/README.md
@@ -384,11 +384,28 @@ _Returns_
 
 Undocumented declaration.
 
+### EditableBlockBindingsPanelItem
+
+Internal dependencies
+
 ### FontSizePicker
 
 _Related_
 
 -   <https://github.com/WordPress/gutenberg/blob/HEAD/packages/block-editor/src/components/font-sizes/README.md>
+
+### getAttributeType
+
+Get the normalized attribute type for block bindings. Converts 'rich-text' to 'string' since rich-text is stored as string.
+
+_Parameters_
+
+-   _blockName_ `string`: The block name.
+-   _attribute_ `string`: The attribute name.
+
+_Returns_
+
+-   `string`: The normalized attribute type.
 
 ### getColorClassName
 
@@ -760,6 +777,10 @@ _Returns_
 ### privateApis
 
 Private @wordpress/block-editor APIs.
+
+### ReadOnlyBlockBindingsPanelItem
+
+Internal dependencies
 
 ### RecursionProvider
