+++ packages/block-editor/src/components/block-bindings/index.js
@@ -1,0 +1,8 @@
+/**
+ * Internal dependencies
+ */
+export {
+	EditableBlockBindingsPanelItem,
+	ReadOnlyBlockBindingsPanelItem,
+} from './attribute';
+export { getAttributeType } from './panel-menu-content';
